Packing function almost complete

+++ bitmap.h
@@ -7,7 +7,7 @@ typedef struct _bmp {
   uint16_t Height;
   uint32_t HeadSize;
   char *Data;
-  uint32_t RealSize;
+  size_t RealSize;
 } bmp_t;
 
 void ReadMemBmp(bmp_t *Bitmap, FILE *FilePointer, size_t Offset);
